Updated theme aditional pages
