+++ lib/widgets/activity_selector_item.dart
@@ -1,6 +1,5 @@
 import 'package:final_project/objects/app_state.dart';
 import 'package:final_project/objects/activity.dart';
-import 'package:final_project/objects/theme.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
