+++ test/objects/group_test.dart
@@ -11,36 +11,4 @@ void main() {
 
     expect(test.abbrev(), 'T');
   });
-  testWidgets('GroupItem has name', (tester) async {
-    Group test = const Group(name: "Test", color: Color(0xFF0F8644), age: 1);
-
-    await tester.pumpWidget(
-        MaterialApp(home: GroupItem(group: test, onTap: ((group) {}))));
-
-    final nameFinder = find.text(test.name);
-
-    expect(nameFinder, findsOneWidget);
-  });
-  testWidgets('GroupItem has Icon abbrev', (tester) async {
-    Group test = const Group(name: "Test", color: Color(0xFF0F8644), age: 1);
-
-    await tester.pumpWidget(
-        MaterialApp(home: GroupItem(group: test, onTap: ((group) {}))));
-
-    final nameFinder = find.text(test.abbrev());
-
-    expect(nameFinder, findsOneWidget);
-  });
-  testWidgets('GroupItem calls onTap callback', (tester) async {
-    Group test = const Group(name: "Test", color: Color(0xFF0F8644), age: 1);
-
-    await tester.pumpWidget(MaterialApp(
-        home: GroupItem(
-            group: test,
-            onTap: ((group) {
-              expect(group, test);
-            }))));
-
-    await tester.tap(find.byType(ListTile));
-  });
 }
